vision_assignment/requirements.txt
absl-py==1.4.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.1.0
arrow==1.2.3
astor==0.8.1
attrs==21.2.0
autopep8==2.0.1
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brotli==1.0.9
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.1.1
command-not-found==0.3
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.4
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
entrypoints==0.4
fastjsonschema==2.16.2
filelock==3.10.7
fqdn==1.5.1
fs==2.4.12
gast==0.5.2
graphviz==0.20.1
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.4
imageio==2.27.0
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
isoduration==20.11.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
json5==0.9.10
jsonpointer==2.3
jsonschema==4.17.3
jupyter-console==6.4.0
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.5
jupyter_server_terminals==0.4.3
jupyterlab==3.5.2
jupyterlab-pygments==0.1.2
jupyterlab_server==2.16.6
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
lz4==3.1.3+dfsg
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==2.0.4
more-itertools==8.10.0
mpmath==0.0.0
nbclassic==0.4.8
nbclient==0.5.6
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.4
netifaces==0.11.0
notebook==6.4.8
notebook_shim==0.2.2
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandocfilters==1.5.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.6.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==4.22.3
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.10.0
pydot==1.4.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.1
PyYAML==5.4.1
pyzmq==24.0.1
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
SecretStorage==3.3.1
Send2Trash==1.8.1b0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
sympy==1.9
systemd-python==234
tensorflow-docs @ git+https://github.com/tensorflow/docs@84bd7bea0c468667dd0ec273d426ed50f23e1a15
tensorflow-hub==0.13.0
terminado==0.13.1
testpath==0.5.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.8.0
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.2.0
urllib3==1.26.13
virtualenv==20.21.0
wadllib==1.3.6
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==2.0.0
zipp==1.0.0
